``dfocus`` now classified
Instead of a collection of routines ported from IDL, the ``dfocus``
tool is now in a class structure to allow the sharing of information
between methods instead of having to pass a whole bunch of stuff back
and forth.

This still needs refinement, as well as rework of some of the more
general IDL functions into the simplified specific case used here.

	modified:   obstools/dfocus.py